+++ configuration/src/main/java/com/synopsys/integration/configuration/property/types/bool/BooleanProperty.java
@@ -3,12 +3,16 @@ package com.synopsys.integration.configuration.property.types.bool;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import com.synopsys.integration.configuration.property.PropertyBuilder;
 import com.synopsys.integration.configuration.property.base.ValuedAlikeProperty;
-import com.synopsys.integration.configuration.property.base.ValuedProperty;
 
 public class BooleanProperty extends ValuedAlikeProperty<Boolean> {
     public BooleanProperty(@NotNull String key, @NotNull Boolean defaultValue) {
         super(key, new BooleanValueParser(), defaultValue);
+    }
+
+    public static PropertyBuilder<BooleanProperty> newBuilder(@NotNull String key, @NotNull Boolean defaultValue) {
+        return new PropertyBuilder<BooleanProperty>().setCreator(() -> new BooleanProperty(key, defaultValue));
     }
 
     @Nullable