+++ configuration/src/main/java/com/synopsys/integration/configuration/property/types/enumallnone/property/AllNoneEnumListProperty.java
@@ -5,9 +5,9 @@ import java.util.List;
 
 import org.jetbrains.annotations.NotNull;
 
+import com.synopsys.integration.configuration.property.PropertyBuilder;
 import com.synopsys.integration.configuration.property.types.enumallnone.enumeration.AllNoneEnum;
 import com.synopsys.integration.configuration.property.types.enumallnone.list.AllNoneEnumList;
-import com.synopsys.integration.configuration.property.types.enumextended.ExtendedEnumListProperty;
 import com.synopsys.integration.configuration.property.types.enumextended.ExtendedEnumListPropertyBase;
 import com.synopsys.integration.configuration.property.types.enumextended.ExtendedEnumValue;
 
@@ -24,6 +24,18 @@ public class AllNoneEnumListProperty<B extends Enum<B>> extends ExtendedEnumList
         super(key, Collections.singletonList(ExtendedEnumValue.ofBaseValue(extendedValue)), AllNoneEnum.class, eClass);
     }
 
+    public static <B extends Enum<B>> PropertyBuilder<AllNoneEnumListProperty<B>> newBuilder(@NotNull String key, List<ExtendedEnumValue<AllNoneEnum, B>> defaultValue, @NotNull Class<B> eClass) {
+        return new PropertyBuilder<AllNoneEnumListProperty<B>>().setCreator(() -> new AllNoneEnumListProperty<>(key, defaultValue, eClass));
+    }
+
+    public static <B extends Enum<B>> PropertyBuilder<AllNoneEnumListProperty<B>> newBuilder(@NotNull String key, @NotNull AllNoneEnum allValue, @NotNull Class<B> eClass) {
+        return new PropertyBuilder<AllNoneEnumListProperty<B>>().setCreator(() -> new AllNoneEnumListProperty<>(key, allValue, eClass));
+    }
+
+    public static <B extends Enum<B>> PropertyBuilder<AllNoneEnumListProperty<B>> newBuilder(@NotNull String key, @NotNull B defaultValue, @NotNull Class<B> eClass) {
+        return new PropertyBuilder<AllNoneEnumListProperty<B>>().setCreator(() -> new AllNoneEnumListProperty<>(key, defaultValue, eClass));
+    }
+
     public AllNoneEnumList<B> toList(List<ExtendedEnumValue<AllNoneEnum, B>> values) {
         return new AllNoneEnumList<>(values, bClass);
     }
@@ -32,4 +44,10 @@ public class AllNoneEnumListProperty<B extends Enum<B>> extends ExtendedEnumList
     public @NotNull AllNoneEnumList<B> convertValue(List<ExtendedEnumValue<AllNoneEnum, B>> value) {
         return new AllNoneEnumList<>(value, bClass);
     }
+
+    public AllNoneEnumListProperty<B> deprecateNone(String reason) {
+        deprecateExtendedValue(AllNoneEnum.NONE, reason);
+        return this;
+    }
+
 }