+++ configuration/src/main/java/com/synopsys/integration/configuration/property/types/enums/EnumListProperty.java
@@ -6,8 +6,8 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import com.synopsys.integration.configuration.parse.ListValueParser;
+import com.synopsys.integration.configuration.property.PropertyBuilder;
 import com.synopsys.integration.configuration.property.base.ValuedAlikeListProperty;
-import com.synopsys.integration.configuration.property.base.ValuedListProperty;
 import com.synopsys.integration.configuration.util.EnumPropertyUtils;
 import com.synopsys.integration.configuration.util.PropertyUtils;
 
@@ -18,6 +18,10 @@ public class EnumListProperty<E extends Enum<E>> extends ValuedAlikeListProperty
     public EnumListProperty(@NotNull String key, @NotNull List<E> defaultValue, @NotNull Class<E> enumClass) {
         super(key, new ListValueParser<>(new EnumValueParser<>(enumClass)), defaultValue);
         this.enumClass = enumClass;
+    }
+
+    public static <E extends Enum<E>> PropertyBuilder<EnumListProperty<E>> newBuilder(@NotNull String key, @NotNull List<E> defaultValue, @NotNull Class<E> enumClass) {
+        return new PropertyBuilder<EnumListProperty<E>>().setCreator(() -> new EnumListProperty<>(key, defaultValue, enumClass));
     }
 
     @Nullable