+++ configuration/src/main/java/com/synopsys/integration/configuration/property/types/integer/NullableIntegerProperty.java
@@ -3,12 +3,16 @@ package com.synopsys.integration.configuration.property.types.integer;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import com.synopsys.integration.configuration.property.PropertyBuilder;
 import com.synopsys.integration.configuration.property.base.NullableAlikeProperty;
-import com.synopsys.integration.configuration.property.base.NullableProperty;
 
 public class NullableIntegerProperty extends NullableAlikeProperty<Integer> {
     public NullableIntegerProperty(@NotNull String key) {
         super(key, new IntegerValueParser());
+    }
+
+    public static PropertyBuilder<NullableIntegerProperty> newBuilder(@NotNull String key) {
+        return new PropertyBuilder<NullableIntegerProperty>().setCreator(() -> new NullableIntegerProperty(key));
     }
 
     @Nullable