+++ configuration/src/main/java/com/synopsys/integration/configuration/property/types/longs/LongProperty.java
@@ -3,12 +3,16 @@ package com.synopsys.integration.configuration.property.types.longs;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import com.synopsys.integration.configuration.property.PropertyBuilder;
 import com.synopsys.integration.configuration.property.base.ValuedAlikeProperty;
-import com.synopsys.integration.configuration.property.base.ValuedProperty;
 
 public class LongProperty extends ValuedAlikeProperty<Long> {
-    public LongProperty(@NotNull String key, Long defaultValue) {
+    public LongProperty(@NotNull String key, @NotNull Long defaultValue) {
         super(key, new LongValueParser(), defaultValue);
+    }
+
+    public static PropertyBuilder<LongProperty> newBuilder(@NotNull String key, @NotNull Long defaultValue) {
+        return new PropertyBuilder<LongProperty>().setCreator(() -> new LongProperty(key, defaultValue));
     }
 
     @Nullable