+++ configuration/src/main/java/com/synopsys/integration/configuration/property/types/string/NullableStringProperty.java
@@ -3,12 +3,16 @@ package com.synopsys.integration.configuration.property.types.string;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import com.synopsys.integration.configuration.property.PropertyBuilder;
 import com.synopsys.integration.configuration.property.base.NullableAlikeProperty;
-import com.synopsys.integration.configuration.property.base.NullableProperty;
 
 public class NullableStringProperty extends NullableAlikeProperty<String> {
     public NullableStringProperty(@NotNull String key) {
         super(key, new StringValueParser());
+    }
+
+    public static PropertyBuilder<NullableStringProperty> newBuilder(@NotNull String key) {
+        return new PropertyBuilder<NullableStringProperty>().setCreator(() -> new NullableStringProperty(key));
     }
 
     @Nullable