+++ src/test/java/com/synopsys/integration/detect/lifecycle/shutdown/ShutdownDeciderTest.java
@@ -40,14 +40,14 @@ public class ShutdownDeciderTest {
 
     @Test
     public void shouldPreserveScanIfDryRun() {
-        PropertyConfiguration configuration = ConfigTestUtils.configOf(Pair.of(DetectProperties.DETECT_BLACKDUCK_SIGNATURE_SCANNER_DRY_RUN.getProperty().getKey(), "true"));
+        PropertyConfiguration configuration = ConfigTestUtils.configOf(Pair.of(DetectProperties.DETECT_BLACKDUCK_SIGNATURE_SCANNER_DRY_RUN.getKey(), "true"));
         CleanupDecision decision = new ShutdownDecider().decideCleanup(configuration, null, null);
         assertTrue(decision.shouldPreserveScan());
     }
 
     @Test
     public void shouldSkipCleanupIfProvidedFalse() {
-        PropertyConfiguration configuration = ConfigTestUtils.configOf(Pair.of(DetectProperties.DETECT_CLEANUP.getProperty().getKey(), "false"));
+        PropertyConfiguration configuration = ConfigTestUtils.configOf(Pair.of(DetectProperties.DETECT_CLEANUP.getKey(), "false"));
         CleanupDecision decision = new ShutdownDecider().decideCleanup(configuration, null, null);
         assertFalse(decision.shouldCleanup());
     }